# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LEAFLINE Plant Store API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /plants:
    get:
      operationId: listPlants
      tags: [plants]
      summary: List plants with optional filtering
      parameters:
        - name: category
          in: query
          schema: { type: string }
        - name: difficulty
          in: query
          schema: { type: string }
        - name: petSafe
          in: query
          schema: { type: boolean }
        - name: lowLight
          in: query
          schema: { type: boolean }
        - name: airPurifier
          in: query
          schema: { type: boolean }
        - name: indoor
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          schema: { type: string }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [price_asc, price_desc, name, popularity, difficulty]
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
      responses:
        "200":
          description: List of plants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plant"
  /plants/{id}:
    get:
      operationId: getPlant
      tags: [plants]
      summary: Get a plant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Plant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plant"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List user orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order from current cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /wishlist:
    get:
      operationId: getWishlist
      tags: [wishlist]
      summary: Get wishlist
      responses:
        "200":
          description: Wishlist plant IDs
          content:
            application/json:
              schema:
                type: array
                items:
                  type: integer
    post:
      operationId: addToWishlist
      tags: [wishlist]
      summary: Add plant to wishlist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WishlistBody"
      responses:
        "200":
          description: Added
  /wishlist/{plantId}:
    delete:
      operationId: removeFromWishlist
      tags: [wishlist]
      summary: Remove plant from wishlist
      parameters:
        - name: plantId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Removed
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    Plant:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        scientificName:
          type: string
        price:
          type: number
        originalPrice:
          type: number
          nullable: true
        category:
          type: string
        difficulty:
          type: string
        lightRequirement:
          type: string
        wateringFrequency:
          type: string
        petSafe:
          type: boolean
        airPurificationRating:
          type: integer
        indoor:
          type: boolean
        description:
          type: string
        careInstructions:
          type: string
        recommendedPlacement:
          type: string
        growthSpeed:
          type: string
        humidityRequirement:
          type: string
        emoji:
          type: string
        gradient:
          type: string
        badge:
          type: string
          nullable: true
        popularity:
          type: integer
      required:
        - id
        - name
        - scientificName
        - price
        - category
        - difficulty
        - lightRequirement
        - wateringFrequency
        - petSafe
        - airPurificationRating
        - indoor
        - description
        - careInstructions
        - recommendedPlacement
        - growthSpeed
        - humidityRequirement
        - emoji
        - gradient
        - popularity
...
    Order:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        totalAmount:
          type: number
        shippingName:
          type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingZip:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
      required: [id, status, totalAmount, shippingName, shippingAddress, shippingCity, shippingZip, items, createdAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        plantId:
          type: integer
        plantName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required: [id, plantId, plantName, quantity, price]
    CreateOrderBody:
      type: object
      properties:
        shippingName:
          type: string
        shippingAddress:
          type: string
        shippingCity:
          type: string
        shippingZip:
          type: string
        paymentMethod:
          type: string
      required: [shippingName, shippingAddress, shippingCity, shippingZip, paymentMethod]
...
    WishlistBody:
      type: object
      properties:
        plantId:
          type: integer
      required: [plantId]
...